pyproject.toml
[project]
name = "datadivr"
version = "0.0.1"
description = "datadivr backend and toolset."
authors = [{ name = "phfr", email = "philipp@lbi-netmed.com" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi>=0.115.4",
    "prompt-toolkit>=3.0.48",
    "rich>=13.9.4",
    "typer>=0.13.0",
    "uvicorn>=0.32.0",
    "websockets>=13.1",
    "pydantic>=2.0.0",
    "structlog>=23.1.0",
    "numpy>=2.0.2",
    "orjson>=3.10.11",
    "plotly>=5.24.1",
    "pillow>=11.0.0",
]

[project.urls]
Homepage = "https://menchelab.github.io/datadivr/"
Repository = "https://github.com/menchelab/datadivr"
Documentation = "https://menchelab.github.io/datadivr/"

[tool.uv]
dev-dependencies = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.20.0",
    "mypy>=0.991",
    "pytest-cov>=4.0.0",
    "ruff>=0.6.9",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "mkdocs-include-markdown-plugin>=7.0.0",
    "httpx>=0.27.2",
    "pytest-asyncio>=0.24.0",
    "psutil>=6.1.0",
    "pytest-xdist>=3.6.1",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["datadivr"]

[tool.mypy]
files = ["datadivr"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
ignore_missing_imports = true
disable_error_code = ["no-any-unimported"]


[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["datadivr"]

[project.scripts]
datadivr = "datadivr.cli:app_cli"
